perf(home): batch trending and watchlist catalog lookups
loadTrending issued up to ~200 sequential FetchDescriptor queries on
the main context (one per TMDB trending title) every time the Home
screen loaded, and loadWatchlist did the same per Trakt watchlist
item. Replace the per-title fetchMovie/fetchSeries helpers with two
batched contains-queries keyed by tmdbId.

The trending/watchlist section moves to HomeView+Trending.swift to
keep HomeView under the file-length cap; the members it drives become
internal, and the file-private LoadState becomes HomeLoadState.
